feat: 컴포넌트 로직 수정 - organisms/Contents 컴포넌트 추가로 파일 재사용성
